dataset-financing-infos/requirements.txt
requests>=2.31.0
feedparser>=6.0.10
pydantic>=2.0.0
python-dateutil>=2.8.2
beautifulsoup4>=4.12.0
pyyaml>=6.0.1
pytest>=7.4.0
tqdm>=4.66.0
click>=8.1.7
